Separate flows.js into subfiles
