+++ src/constants/flows/walkthrough.js
@@ -1,204 +1,17 @@
-const { getRole, getChannel } = require("./resolvers.js");
-
-module.exports = {
-  limitTriggers: 1,
-  limitActions: 3,
-  limitFlows: 5
-};
-
-module.exports.propertyTypes = {
-  "numberX": {
-    "short": "Number (X)",
-    "help": "This can be any positive number.",
-    "convert": async num => parseInt(num) || null
-  },
-  "regex": {
-    "short": "Regex",
-    "help": "Get help on how to create a regex here: https://flaviocopes.com/javascript-regular-expressions/#regular-expressions-choices",
-    "convert": async regex => {
-      try {
-        new RegExp(regex);
-        return regex;
-      } catch(e) {
-        return false;
-      }
-    }
-  },
-  "role": {
-    "short": "Role",
-    "help": "Any role. Make sure the role is below Countr's highest role.",
-    "convert": async (search, { guild }) => {
-      const result = await getRole(search, guild);
-      if (result && result.id !== guild.id) return result.id; else return null;
-    },
-    "format": async roleid => `<@&${roleid}>`
-  },
-  "channel": {
-    "short": "Channel",
-    "help": "Any channel. Make sure Countr has access to the channel, and that it is a text based channel. (news channels also work)",
-    "convert": async (search, { guild }) => {
-      const result = await getChannel(search, guild);
-      if (result) return result.id; else return null;
-    },
-    "format": async channelid => `<#${channelid}>`
-  },
-  "text": {
-    "short": "Text",
-    "help": [
-      "Any text. Get creative with these placeholders:",
-      "• `{count}` The count that triggered this flow",
-      "• `{mention}` Mentions the user who triggered this flow",
-      "• `{tag}` The tag of the user who triggered this flow",
-      "• `{username}` The username of the user who triggered this flow",
-      "• `{nickname}` The nickname of the user who triggered this flow",
-      "• `{everyone}` Mentions the everyone-role",
-      "• `{score}` The new score of the user who triggered this flow"
-    ].join("\n")
-  }
-};
-
-for (const i in module.exports.propertyTypes) module.exports.propertyTypes[i] = Object.assign({
-  "short": "N/A",
-  "help": null,
-  "convert": async any => any,
-  "format": async any => any
-}, module.exports.propertyTypes[i]);
-
-module.exports.flow = {
-  triggers: {
-    "each": {
-      "short": "Each X number",
-      "long": "This will get triggered whenever a user counts a multiple of X. For example, if X is 10, this will trigger on 10, 20, 30 etc.",
-      "properties": [
-        module.exports.propertyTypes.numberX
-      ],
-      "explanation": "When someone counts a multiplication of {0}",
-      "check": async ({ count }, [ number ]) => count % number == 0
-    },
-    "only": {
-      "short": "Only number X",
-      "long": "This will get triggered whenever a user counts the number X, and only the number X.",
-      "properties": [
-        module.exports.propertyTypes.numberX
-      ],
-      "explanation": "When someone counts the number {0}",
-      "check": async ({ count }, [ number ]) => count == number
-    },
-    "score": {
-      "short": "Score of X",
-      "long": "This will get triggered whenever a user has counted a total of X counts.",
-      "properties": [
-        module.exports.propertyTypes.numberX
-      ],
-      "explanation": "When someone gets a score of {0}",
-      "check": async ({ score }, [ number ]) => score == number
-    },
-    "regex": {
-      "short": "Regex match",
-      "long": "This will get triggered when a count matches a regex.",
-      "properties": [
-        module.exports.propertyTypes.regex
-      ],
-      "explanation": "When a message matches the regex `{0}`",
-      "check": async ({ message: { content } }, [ regex ]) => (new RegExp(regex)).test(content)
-    }
-  },
-  actions: {
-    "giverole": {
-      "short": "Give a role to the user",
-      "long": "This will add a role to the user who triggered this flow.",
-      "properties": [
-        module.exports.propertyTypes.role
-      ],
-      "explanation": "Add the user to {0}",
-      "run": async ({ message: { guild } }, [ roleID ]) => {
-        const role = guild.roles.resolve(roleID);
-        if (role) await Promise.all(role.members.map(async member => await member.roles.remove(roleID).catch()));
-      }
-    },
-    "takerole": {
-      "short": "Take a role away from the user",
-      "long": "This will remove a role from the user who triggered this flow.",
-      "properties": [
-        module.exports.propertyTypes.role
-      ],
-      "explanation": "Remove the user from {0}",
-      "run": async ({ message: { member } }, [ roleID ]) => await member.roles.remove(roleID).catch()
-    },
-    "prunerole": {
-      "short": "Remove everyone from a role",
-      "long": "This will remove everyone from this role.",
-      "properties": [
-        module.exports.propertyTypes.role
-      ],
-      "explanation": "Remove everyone from {0}",
-      "run": async ({ message: { member } }, [ roleID ]) => await member.roles.add(roleID).catch()
-    },
-    "pin": {
-      "short": "Pin the count message",
-      "explanation": "Pin the count",
-      "run": async ({ message }) => await message.pin()
-    },
-    "sendmessage": {
-      "short": "Send a message",
-      "long": "This will send a message in a channel (it doesn't have to be the counting channel!)",
-      "properties": [
-        module.exports.propertyTypes.channel,
-        module.exports.propertyTypes.text
-      ],
-      "explanation": "Send a message in {0}: ```{1}```",
-      "run": async ({ count, score, message: { guild, member, author } }, [ channelID, text ]) => {
-        const channel = guild.channels.resolve(channelID);
-        if (channel) await channel.send(text
-          .replace(/{count}/gi, count)
-          .replace(/{mention}/gi, member.toString())
-          .replace(/{tag}/gi, author.tag)
-          .replace(/{username}/gi, author.username)
-          .replace(/{nickname}/gi, member.displayName)
-          .replace(/{everyone}/gi, guild.roles.everyone.toString())
-          .replace(/{score}/gi, score)
-        , {
-          disableMentions: "none"
-        }).catch();
-      }
-    },
-    "lock": {
-      "short": "Lock the counting channel",
-      "long": "This will lock the counting channel for the everyone-role, and will be read-only.",
-      "explanation": "Lock the counting channel",
-      "run": async ({ message: { channel, guild } }) => await channel.updateOverwrite(guild.roles.everyone, { SEND_MESSAGES: false }).catch()
-    }
-  }
-};
-
-for (const i in module.exports.flow.triggers) module.exports.flow.triggers[i] = Object.assign({
-  "short": "N/A",
-  "long": null,
-  "properties": [],
-  "explanation": "N/A",
-  "check": async any => any
-}, module.exports.flow.triggers[i]);
-
-for (const i in module.exports.flow.actions) module.exports.flow.actions[i] = Object.assign({
-  "short": "N/A",
-  "long": null,
-  "properties": [],
-  "explanation": "N/A",
-  "run": async () => null
-}, module.exports.flow.actions[i]);
+const { flow } = require("./flow.js");
 
 module.exports.formatExplanation = async ({ type, data }) => {
-  let { properties, explanation } = module.exports.flow.triggers[type] || module.exports.flow.actions[type];
+  let { properties, explanation } = flow.triggers[type] || flow.actions[type];
   for (const i in properties) explanation = explanation.replace(`{${i}}`, await properties[i].format(data[i]));
   return explanation;
 };
 
 const
-  allActionTypes = Object.keys(module.exports.flow.actions),
-  allActions = Object.values(module.exports.flow.actions),
-  allTriggerTypes = Object.keys(module.exports.flow.triggers),
-  allTriggers = Object.values(module.exports.flow.triggers),
-  config = require("../../config.json");
+  allActionTypes = Object.keys(flow.actions),
+  allActions = Object.values(flow.actions),
+  allTriggerTypes = Object.keys(flow.triggers),
+  allTriggers = Object.values(flow.triggers),
+  config = require("../../../config.json");
 
 module.exports.flowWalkthrough = async (guild, author, channel, newFlow, generateEmbed, pinned) => {
   // add slots to fill it up when editing
